Move match-related code to a file
